Moved modulator functionality to new class and added support for getting modulator values
