+++ Assets/Scripts/Projectile/Modulator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: df45877e7a2bd1442b0d523eae31001e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
