+++ Assets/Scripts/Projectile/ProjectileMovement.cs
@@ -12,30 +12,6 @@ public enum MovementActionType
     InstantRotateTowardsPlayer
 }
 
-[System.Serializable]
-public struct ModulatorProperties
-{
-    public float amplitude;
-    public float frequency;
-    public float lift;
-    public float offset;
-}
-
-public enum RotationModulationType
-{
-    Sin,
-    Cos
-}
-
-[System.Serializable]
-public struct ModulatorEntry
-{
-    public RotationModulationType modulationType;
-    public ModulatorProperties waveProperties;
-}
-
-
-
 namespace Projectile
 {
     public class ProjectileMovement : MonoBehaviour
@@ -56,7 +32,7 @@ namespace Projectile
         MovementEventActions eventActions = new MovementEventActions();
 
         [SerializeField]
-        private List<ModulatorEntry> modulatorFunctionsList;
+        private List<Modulator.ModulatorEntry> modulatorFunctionsList;
 
 
         private void Start()
@@ -110,16 +86,6 @@ namespace Projectile
 
 
             return false;
-        }
-
-        private void ApplyRotationModulators()
-        {
-            return;
-        }
-
-        private void CalculateModulationStrength()
-        {
-
         }
 
         private bool CheckEvents()
